Fix for http://java.net/jira/browse/GRIZZLY-1109 (Long type for max-swallowing-input-bytes fails validation when using asadmin set.)

+++ modules/config/src/main/java/com/sun/grizzly/config/dom/Http.java
@@ -204,7 +204,7 @@ public interface Http extends ConfigBeanProxy, Injectable, PropertyBag {
 
     void setMaxPostSizeBytes(String max);
 
-    @Attribute(defaultValue = "" + MAX_SWALLOWING_INPUT_BYTES, dataType = Long.class)
+    @Attribute(defaultValue = "" + MAX_SWALLOWING_INPUT_BYTES, dataType = Integer.class)
     String getMaxSwallowingInputBytes();
 
     void setMaxSwallowingInputBytes(String max);
